requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0
loguru==0.7.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.27.0
pytest==8.0.0
pdf2docx==0.5.8
pymupdf==1.23.21
pillow==10.2.0
python-docx==1.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5
